Collapse PhoneNumberUtilTest ports into one file mirroring Java order
Merge the topically-split phonenumberutil_*_test.go ports back into a single
phonenumberutil_test.go whose test functions are ordered to match upstream's
PhoneNumberUtilTest.java method-by-method, so the file can be diffed straight
against upstream when reconciling. The split was only ever justified as a
staging area while the ad-hoc real-metadata tests were migrated to faithful
synthetic-metadata ports; that migration is effectively complete, so the split
now adds drift risk (which file does a new upstream test belong in?) without
benefit.

Also rename the former ad-hoc test file to phonenumberutil_internal_test.go and
rewrite its header: what remains has no upstream PhoneNumberUtilTest counterpart
(direct tests of internal helpers and the RegexCache strictness behaviour), so
the "migration in progress / shrink to nothing" framing was stale.

No test logic changes: the merged file contains exactly the same 120 test
functions as the 11 files it replaces (verified by set diff), and the full
suite passes.
