--- phonenumberutil_types_test.go
@@ -1,240 +1,0 @@
-package phonenumbers
-
-// Faithful port of upstream libphonenumber's PhoneNumberUtilTest.java tests for
-// getSupportedTypesForRegion/NonGeoEntity, getInvalidExampleNumber, and
-// isPossibleNumberForType(WithReason), run against the synthetic test metadata.
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-	"google.golang.org/protobuf/proto"
-)
-
-func TestGetSupportedTypesForRegion(t *testing.T) {
-	useTestMetadata(t)
-	assert.True(t, GetSupportedTypesForRegion(regionCode.BR)[FIXED_LINE])
-	// Our test data has no mobile numbers for Brazil.
-	assert.False(t, GetSupportedTypesForRegion(regionCode.BR)[MOBILE])
-	// UNKNOWN should never be returned.
-	assert.False(t, GetSupportedTypesForRegion(regionCode.BR)[UNKNOWN])
-	// In the US, many numbers are classified as FIXED_LINE_OR_MOBILE; but we don't want to expose
-	// this as a supported type, instead we say FIXED_LINE and MOBILE are both present.
-	assert.True(t, GetSupportedTypesForRegion(regionCode.US)[FIXED_LINE])
-	assert.True(t, GetSupportedTypesForRegion(regionCode.US)[MOBILE])
-	assert.False(t, GetSupportedTypesForRegion(regionCode.US)[FIXED_LINE_OR_MOBILE])
-
-	// Test the invalid region code.
-	assert.Equal(t, 0, len(GetSupportedTypesForRegion(regionCode.ZZ)))
-}
-
-func TestGetSupportedTypesForNonGeoEntity(t *testing.T) {
-	useTestMetadata(t)
-	// No data exists for 999 at all, no types should be returned.
-	assert.Equal(t, 0, len(GetSupportedTypesForNonGeoEntity(999)))
-
-	typesFor979 := GetSupportedTypesForNonGeoEntity(979)
-	assert.True(t, typesFor979[PREMIUM_RATE])
-	assert.False(t, typesFor979[MOBILE])
-	assert.False(t, typesFor979[UNKNOWN])
-}
-
-func TestGetInvalidExampleNumber(t *testing.T) {
-	useTestMetadata(t)
-	// RegionCode 001 is reserved for supporting non-geographical country calling codes.
-	assert.Nil(t, GetInvalidExampleNumber(regionCode.UN001))
-	assert.Nil(t, GetInvalidExampleNumber("CS"))
-	usInvalidNumber := GetInvalidExampleNumber(regionCode.US)
-	assert.Equal(t, int32(1), usInvalidNumber.GetCountryCode())
-	assert.NotEqual(t, uint64(0), usInvalidNumber.GetNationalNumber())
-}
-
-func TestIsPossibleNumberForTypeDifferentTypeLengths(t *testing.T) {
-	useTestMetadata(t)
-	// We use Argentinian numbers since they have different possible lengths for different types.
-	number := &PhoneNumber{CountryCode: proto.Int32(54), NationalNumber: proto.Uint64(12345)}
-	// Too short for any Argentinian number, including fixed-line.
-	assert.False(t, IsPossibleNumberForType(number, FIXED_LINE))
-	assert.False(t, IsPossibleNumberForType(number, UNKNOWN))
-
-	// 6-digit numbers are okay for fixed-line.
-	number.NationalNumber = proto.Uint64(123456)
-	assert.True(t, IsPossibleNumberForType(number, UNKNOWN))
-	assert.True(t, IsPossibleNumberForType(number, FIXED_LINE))
-	// But too short for mobile.
-	assert.False(t, IsPossibleNumberForType(number, MOBILE))
-	// And too short for toll-free.
-	assert.False(t, IsPossibleNumberForType(number, TOLL_FREE))
-
-	// The same applies to 9-digit numbers.
-	number.NationalNumber = proto.Uint64(123456789)
-	assert.True(t, IsPossibleNumberForType(number, UNKNOWN))
-	assert.True(t, IsPossibleNumberForType(number, FIXED_LINE))
-	assert.False(t, IsPossibleNumberForType(number, MOBILE))
-	assert.False(t, IsPossibleNumberForType(number, TOLL_FREE))
-
-	// 10-digit numbers are universally possible.
-	number.NationalNumber = proto.Uint64(1234567890)
-	assert.True(t, IsPossibleNumberForType(number, UNKNOWN))
-	assert.True(t, IsPossibleNumberForType(number, FIXED_LINE))
-	assert.True(t, IsPossibleNumberForType(number, MOBILE))
-	assert.True(t, IsPossibleNumberForType(number, TOLL_FREE))
-
-	// 11-digit numbers are only possible for mobile numbers.
-	number.NationalNumber = proto.Uint64(12345678901)
-	assert.True(t, IsPossibleNumberForType(number, UNKNOWN))
-	assert.False(t, IsPossibleNumberForType(number, FIXED_LINE))
-	assert.True(t, IsPossibleNumberForType(number, MOBILE))
-	assert.False(t, IsPossibleNumberForType(number, TOLL_FREE))
-}
-
-func TestIsPossibleNumberForTypeLocalOnly(t *testing.T) {
-	useTestMetadata(t)
-	number := &PhoneNumber{CountryCode: proto.Int32(49), NationalNumber: proto.Uint64(12)}
-	// Here we test a number length which matches a local-only length.
-	assert.True(t, IsPossibleNumberForType(number, UNKNOWN))
-	assert.True(t, IsPossibleNumberForType(number, FIXED_LINE))
-	// Mobile numbers must be 10 or 11 digits, and there are no local-only lengths.
-	assert.False(t, IsPossibleNumberForType(number, MOBILE))
-}
-
-func TestIsPossibleNumberForTypeDataMissingForSizeReasons(t *testing.T) {
-	useTestMetadata(t)
-	number := &PhoneNumber{CountryCode: proto.Int32(55), NationalNumber: proto.Uint64(12345678)}
-	// Local-only number.
-	assert.True(t, IsPossibleNumberForType(number, UNKNOWN))
-	assert.True(t, IsPossibleNumberForType(number, FIXED_LINE))
-
-	number.NationalNumber = proto.Uint64(1234567890)
-	assert.True(t, IsPossibleNumberForType(number, UNKNOWN))
-	assert.True(t, IsPossibleNumberForType(number, FIXED_LINE))
-}
-
-func TestIsPossibleNumberForTypeNumberTypeNotSupportedForRegion(t *testing.T) {
-	useTestMetadata(t)
-	number := &PhoneNumber{CountryCode: proto.Int32(55), NationalNumber: proto.Uint64(12345678)}
-	// There are *no* mobile numbers for this region at all, so we return false.
-	assert.False(t, IsPossibleNumberForType(number, MOBILE))
-	// This matches a fixed-line length though.
-	assert.True(t, IsPossibleNumberForType(number, FIXED_LINE))
-	assert.True(t, IsPossibleNumberForType(number, FIXED_LINE_OR_MOBILE))
-
-	// There are *no* fixed-line OR mobile numbers for this country calling code at all.
-	number = &PhoneNumber{CountryCode: proto.Int32(979), NationalNumber: proto.Uint64(123456789)}
-	assert.False(t, IsPossibleNumberForType(number, MOBILE))
-	assert.False(t, IsPossibleNumberForType(number, FIXED_LINE))
-	assert.False(t, IsPossibleNumberForType(number, FIXED_LINE_OR_MOBILE))
-	assert.True(t, IsPossibleNumberForType(number, PREMIUM_RATE))
-}
-
-func TestIsPossibleNumberForTypeWithReasonDifferentTypeLengths(t *testing.T) {
-	useTestMetadata(t)
-	number := &PhoneNumber{CountryCode: proto.Int32(54), NationalNumber: proto.Uint64(12345)}
-	// Too short for any Argentinian number.
-	assert.Equal(t, TOO_SHORT, IsPossibleNumberForTypeWithReason(number, UNKNOWN))
-	assert.Equal(t, TOO_SHORT, IsPossibleNumberForTypeWithReason(number, FIXED_LINE))
-
-	// 6-digit numbers are okay for fixed-line.
-	number.NationalNumber = proto.Uint64(123456)
-	assert.Equal(t, IS_POSSIBLE, IsPossibleNumberForTypeWithReason(number, UNKNOWN))
-	assert.Equal(t, IS_POSSIBLE, IsPossibleNumberForTypeWithReason(number, FIXED_LINE))
-	assert.Equal(t, TOO_SHORT, IsPossibleNumberForTypeWithReason(number, MOBILE))
-	assert.Equal(t, TOO_SHORT, IsPossibleNumberForTypeWithReason(number, TOLL_FREE))
-
-	// The same applies to 9-digit numbers.
-	number.NationalNumber = proto.Uint64(123456789)
-	assert.Equal(t, IS_POSSIBLE, IsPossibleNumberForTypeWithReason(number, UNKNOWN))
-	assert.Equal(t, IS_POSSIBLE, IsPossibleNumberForTypeWithReason(number, FIXED_LINE))
-	assert.Equal(t, TOO_SHORT, IsPossibleNumberForTypeWithReason(number, MOBILE))
-	assert.Equal(t, TOO_SHORT, IsPossibleNumberForTypeWithReason(number, TOLL_FREE))
-
-	// 10-digit numbers are universally possible.
-	number.NationalNumber = proto.Uint64(1234567890)
-	assert.Equal(t, IS_POSSIBLE, IsPossibleNumberForTypeWithReason(number, UNKNOWN))
-	assert.Equal(t, IS_POSSIBLE, IsPossibleNumberForTypeWithReason(number, FIXED_LINE))
-	assert.Equal(t, IS_POSSIBLE, IsPossibleNumberForTypeWithReason(number, MOBILE))
-	assert.Equal(t, IS_POSSIBLE, IsPossibleNumberForTypeWithReason(number, TOLL_FREE))
-
-	// 11-digit numbers are only possible for mobile numbers.
-	number.NationalNumber = proto.Uint64(12345678901)
-	assert.Equal(t, IS_POSSIBLE, IsPossibleNumberForTypeWithReason(number, UNKNOWN))
-	assert.Equal(t, TOO_LONG, IsPossibleNumberForTypeWithReason(number, FIXED_LINE))
-	assert.Equal(t, IS_POSSIBLE, IsPossibleNumberForTypeWithReason(number, MOBILE))
-	assert.Equal(t, TOO_LONG, IsPossibleNumberForTypeWithReason(number, TOLL_FREE))
-}
-
-func TestIsPossibleNumberForTypeWithReasonLocalOnly(t *testing.T) {
-	useTestMetadata(t)
-	number := &PhoneNumber{CountryCode: proto.Int32(49), NationalNumber: proto.Uint64(12)}
-	assert.Equal(t, IS_POSSIBLE_LOCAL_ONLY, IsPossibleNumberForTypeWithReason(number, UNKNOWN))
-	assert.Equal(t, IS_POSSIBLE_LOCAL_ONLY, IsPossibleNumberForTypeWithReason(number, FIXED_LINE))
-	assert.Equal(t, TOO_SHORT, IsPossibleNumberForTypeWithReason(number, MOBILE))
-}
-
-func TestIsPossibleNumberForTypeWithReasonDataMissingForSizeReasons(t *testing.T) {
-	useTestMetadata(t)
-	number := &PhoneNumber{CountryCode: proto.Int32(55), NationalNumber: proto.Uint64(12345678)}
-	assert.Equal(t, IS_POSSIBLE_LOCAL_ONLY, IsPossibleNumberForTypeWithReason(number, UNKNOWN))
-	assert.Equal(t, IS_POSSIBLE_LOCAL_ONLY, IsPossibleNumberForTypeWithReason(number, FIXED_LINE))
-
-	number.NationalNumber = proto.Uint64(1234567890)
-	assert.Equal(t, IS_POSSIBLE, IsPossibleNumberForTypeWithReason(number, UNKNOWN))
-	assert.Equal(t, IS_POSSIBLE, IsPossibleNumberForTypeWithReason(number, FIXED_LINE))
-}
-
-func TestIsPossibleNumberForTypeWithReasonNumberTypeNotSupportedForRegion(t *testing.T) {
-	useTestMetadata(t)
-	// There are *no* mobile numbers for this region at all, so we return INVALID_LENGTH.
-	number := &PhoneNumber{CountryCode: proto.Int32(55), NationalNumber: proto.Uint64(12345678)}
-	assert.Equal(t, INVALID_LENGTH, IsPossibleNumberForTypeWithReason(number, MOBILE))
-	// This matches a fixed-line length though.
-	assert.Equal(t, IS_POSSIBLE_LOCAL_ONLY, IsPossibleNumberForTypeWithReason(number, FIXED_LINE_OR_MOBILE))
-	// This is too short for fixed-line, and no mobile numbers exist.
-	number = &PhoneNumber{CountryCode: proto.Int32(55), NationalNumber: proto.Uint64(1234567)}
-	assert.Equal(t, INVALID_LENGTH, IsPossibleNumberForTypeWithReason(number, MOBILE))
-	assert.Equal(t, TOO_SHORT, IsPossibleNumberForTypeWithReason(number, FIXED_LINE_OR_MOBILE))
-	assert.Equal(t, TOO_SHORT, IsPossibleNumberForTypeWithReason(number, FIXED_LINE))
-
-	// This is too short for mobile, and no fixed-line numbers exist.
-	number = &PhoneNumber{CountryCode: proto.Int32(882), NationalNumber: proto.Uint64(1234567)}
-	assert.Equal(t, TOO_SHORT, IsPossibleNumberForTypeWithReason(number, MOBILE))
-	assert.Equal(t, TOO_SHORT, IsPossibleNumberForTypeWithReason(number, FIXED_LINE_OR_MOBILE))
-	assert.Equal(t, INVALID_LENGTH, IsPossibleNumberForTypeWithReason(number, FIXED_LINE))
-
-	// There are *no* fixed-line OR mobile numbers for this country calling code at all.
-	number = &PhoneNumber{CountryCode: proto.Int32(979), NationalNumber: proto.Uint64(123456789)}
-	assert.Equal(t, INVALID_LENGTH, IsPossibleNumberForTypeWithReason(number, MOBILE))
-	assert.Equal(t, INVALID_LENGTH, IsPossibleNumberForTypeWithReason(number, FIXED_LINE))
-	assert.Equal(t, INVALID_LENGTH, IsPossibleNumberForTypeWithReason(number, FIXED_LINE_OR_MOBILE))
-	assert.Equal(t, IS_POSSIBLE, IsPossibleNumberForTypeWithReason(number, PREMIUM_RATE))
-}
-
-func TestIsPossibleNumberForTypeWithReasonFixedLineOrMobile(t *testing.T) {
-	useTestMetadata(t)
-	// For FIXED_LINE_OR_MOBILE, a number should be considered valid if it matches the possible
-	// lengths for mobile *or* fixed-line numbers.
-	number := &PhoneNumber{CountryCode: proto.Int32(290), NationalNumber: proto.Uint64(1234)}
-	assert.Equal(t, TOO_SHORT, IsPossibleNumberForTypeWithReason(number, FIXED_LINE))
-	assert.Equal(t, IS_POSSIBLE, IsPossibleNumberForTypeWithReason(number, MOBILE))
-	assert.Equal(t, IS_POSSIBLE, IsPossibleNumberForTypeWithReason(number, FIXED_LINE_OR_MOBILE))
-
-	number.NationalNumber = proto.Uint64(12345)
-	assert.Equal(t, TOO_SHORT, IsPossibleNumberForTypeWithReason(number, FIXED_LINE))
-	assert.Equal(t, TOO_LONG, IsPossibleNumberForTypeWithReason(number, MOBILE))
-	assert.Equal(t, INVALID_LENGTH, IsPossibleNumberForTypeWithReason(number, FIXED_LINE_OR_MOBILE))
-
-	number.NationalNumber = proto.Uint64(123456)
-	assert.Equal(t, IS_POSSIBLE, IsPossibleNumberForTypeWithReason(number, FIXED_LINE))
-	assert.Equal(t, TOO_LONG, IsPossibleNumberForTypeWithReason(number, MOBILE))
-	assert.Equal(t, IS_POSSIBLE, IsPossibleNumberForTypeWithReason(number, FIXED_LINE_OR_MOBILE))
-
-	number.NationalNumber = proto.Uint64(1234567)
-	assert.Equal(t, TOO_LONG, IsPossibleNumberForTypeWithReason(number, FIXED_LINE))
-	assert.Equal(t, TOO_LONG, IsPossibleNumberForTypeWithReason(number, MOBILE))
-	assert.Equal(t, TOO_LONG, IsPossibleNumberForTypeWithReason(number, FIXED_LINE_OR_MOBILE))
-
-	// 8-digit numbers are possible for toll-free and premium-rate numbers only.
-	number.NationalNumber = proto.Uint64(12345678)
-	assert.Equal(t, IS_POSSIBLE, IsPossibleNumberForTypeWithReason(number, TOLL_FREE))
-	assert.Equal(t, TOO_LONG, IsPossibleNumberForTypeWithReason(number, FIXED_LINE_OR_MOBILE))
-}
